- Add ActiveSetAnalysis and move all EncodeSet classes there
