--- dartagnan/src/main/java/com/dat3m/dartagnan/encoding/EncodeSets.java
@@ -1,187 +1,0 @@
-package com.dat3m.dartagnan.encoding;
-
-import com.dat3m.dartagnan.program.event.Event;
-import com.dat3m.dartagnan.verification.Context;
-import com.dat3m.dartagnan.wmm.Constraint.Visitor;
-import com.dat3m.dartagnan.wmm.Definition;
-import com.dat3m.dartagnan.wmm.Relation;
-import com.dat3m.dartagnan.wmm.analysis.RelationAnalysis;
-import com.dat3m.dartagnan.wmm.definition.*;
-import com.dat3m.dartagnan.wmm.utils.graph.EventGraph;
-import com.dat3m.dartagnan.wmm.utils.graph.mutable.MapEventGraph;
-import com.dat3m.dartagnan.wmm.utils.graph.mutable.MutableEventGraph;
-
-import java.util.HashMap;
-import java.util.Map;
-import java.util.Set;
-
-// Propagates relationships in a verification task that need to be constrained in an SMT-based verification.
-final class EncodeSets implements Visitor<Map<Relation, EventGraph>> {
-
-    private final RelationAnalysis ra;
-    EventGraph news;
-
-    EncodeSets(Context analysisContext) {
-        ra = analysisContext.requires(RelationAnalysis.class);
-    }
-
-    @Override
-    public Map<Relation, EventGraph> visitDefinition(Definition def) {
-        return Map.of();
-    }
-
-    @Override
-    public Map<Relation, EventGraph> visitUnion(Union union) {
-        final Map<Relation, EventGraph> m = new HashMap<>();
-        for (Relation r : union.getOperands()) {
-            m.merge(r, filterUnknowns(news, r), EventGraph::union);
-        }
-        return m;
-    }
-
-    @Override
-    public Map<Relation, EventGraph> visitIntersection(Intersection inter) {
-        final Map<Relation, EventGraph> m = new HashMap<>();
-        for (Relation r : inter.getOperands()) {
-            m.merge(r, filterUnknowns(news, r), EventGraph::union);
-        }
-        return m;
-    }
-
-    @Override
-    public Map<Relation, EventGraph> visitDifference(Difference diff) {
-        final Relation r1 = diff.getMinuend();
-        final Relation r2 = diff.getSubtrahend();
-        return map(r1, filterUnknowns(news, r1), r2, filterUnknowns(news, r2));
-    }
-
-    @Override
-    public Map<Relation, EventGraph> visitComposition(Composition comp) {
-        if (news.isEmpty()) {
-            return Map.of();
-        }
-
-        final Relation r1 = comp.getLeftOperand();
-        final Relation r2 = comp.getRightOperand();
-        final MutableEventGraph set1 = new MapEventGraph();
-        final MutableEventGraph set2 = new MapEventGraph();
-        final RelationAnalysis.Knowledge k1 = ra.getKnowledge(r1);
-        final RelationAnalysis.Knowledge k2 = ra.getKnowledge(r2);
-        Map<Event, Set<Event>> out = ra.getKnowledge(r1).getMaySet().getOutMap();
-        news.apply((e1, e2) -> {
-            for (Event e : out.getOrDefault(e1, Set.of())) {
-                if (k2.getMaySet().contains(e, e2)) {
-                    if (!k1.getMustSet().contains(e1, e)) {
-                        set1.add(e1, e);
-                    }
-                    if (!k2.getMustSet().contains(e, e2)) {
-                        set2.add(e, e2);
-                    }
-                }
-            }
-        });
-        return map(r1, set1, r2, set2);
-    }
-
-    @Override
-    public Map<Relation, EventGraph> visitProjection(Projection projection) {
-        final MutableEventGraph result = new MapEventGraph();
-        final RelationAnalysis.Knowledge k1 = ra.getKnowledge(projection.getOperand());
-        final EventGraph mayGraph = k1.getMaySet();
-        final EventGraph mustGraph = k1.getMustSet();
-        final boolean dom = projection.getDimension() == Projection.Dimension.DOMAIN;
-        final Map<Event, Set<Event>> altMap = dom ? mayGraph.getOutMap() : mayGraph.getInMap();
-        news.apply((e1, e2) -> {
-            assert e1.equals(e2);
-            for (Event alt : altMap.getOrDefault(e1, Set.of())) {
-                if (!mustGraph.contains(dom ? e1 : alt, dom ? alt : e1)) {
-                    result.add(dom ? e1 : alt, dom ? alt : e1);
-                }
-            }
-        });
-        return Map.of(projection.getOperand(), result);
-    }
-
-    @Override
-    public Map<Relation, EventGraph> visitSetIdentity(SetIdentity id) {
-        return Map.of(id.getDomain(), filterUnknowns(news, id.getDomain()));
-    }
-
-    @Override
-    public Map<Relation, EventGraph> visitInverse(Inverse inv) {
-        return Map.of(inv.getOperand(), filterUnknowns(news.inverse(), inv.getOperand()));
-    }
-
-    @Override
-    public Map<Relation, EventGraph> visitProduct(CartesianProduct product) {
-        final RelationAnalysis.Knowledge k1 = ra.getKnowledge(product.getDomain());
-        final RelationAnalysis.Knowledge k2 = ra.getKnowledge(product.getRange());
-        final MutableEventGraph set1 = new MapEventGraph();
-        final MutableEventGraph set2 = new MapEventGraph();
-        for (Event e1 : news.getDomain()) {
-            if (k1.getMaySet().contains(e1, e1) && !k1.getMustSet().contains(e1, e1)) {
-                set1.add(e1, e1);
-            }
-            for (Event e2 : news.getRange(e1)) {
-                if (k2.getMaySet().contains(e2, e2) && !k2.getMustSet().contains(e2, e2)) {
-                    set2.add(e2, e2);
-                }
-            }
-        }
-        return map(product.getDomain(), set1, product.getRange(), set2);
-    }
-
-    @Override
-    public Map<Relation, EventGraph> visitTransitiveClosure(TransitiveClosure trans) {
-        final Relation rel = trans.getDefinedRelation();
-        final Relation r1 = trans.getOperand();
-        MutableEventGraph factors = new MapEventGraph();
-        final RelationAnalysis.Knowledge k0 = ra.getKnowledge(rel);
-        Map<Event, Set<Event>> out = k0.getMaySet().getOutMap();
-        news.apply((e1, e2) -> {
-            for (Event e : out.getOrDefault(e1, Set.of())) {
-                if (k0.getMaySet().contains(e, e2)) {
-                    if (!k0.getMustSet().contains(e1, e)) {
-                        factors.add(e1, e);
-                    }
-                    if (!k0.getMustSet().contains(e, e2)) {
-                        factors.add(e, e2);
-                    }
-                }
-            }
-        });
-        return map(rel, factors, r1, filterUnknowns(news, r1));
-    }
-
-    @Override
-    public Map<Relation, EventGraph> visitLinuxCriticalSections(LinuxCriticalSections rscs) {
-        MutableEventGraph queue = new MapEventGraph();
-        final RelationAnalysis.Knowledge k0 = ra.getKnowledge(rscs.getDefinedRelation());
-        Map<Event, Set<Event>> in = k0.getMaySet().getInMap();
-        Map<Event, Set<Event>> out = k0.getMaySet().getOutMap();
-        news.apply((lock, unlock) -> {
-            for (Event e : in.getOrDefault(unlock, Set.of())) {
-                if (lock.getGlobalId() < e.getGlobalId() && e.getGlobalId() < unlock.getGlobalId()) {
-                    queue.add(e, unlock);
-                }
-            }
-            for (Event e : out.getOrDefault(lock, Set.of())) {
-                if (lock.getGlobalId() < e.getGlobalId() && e.getGlobalId() < unlock.getGlobalId()) {
-                    queue.add(lock, e);
-                }
-            }
-        });
-        return Map.of(rscs.getDefinedRelation(), queue);
-    }
-
-    private EventGraph filterUnknowns(EventGraph news, Relation relation) {
-        RelationAnalysis.Knowledge k = ra.getKnowledge(relation);
-        EventGraph may = k.getMaySet();
-        EventGraph must = k.getMustSet();
-        return news.filter((e1, e2) -> may.contains(e1, e2) && !must.contains(e1, e2));
-    }
-
-    private static Map<Relation, EventGraph> map(Relation r1, EventGraph s1, Relation r2, EventGraph s2) {
-        return r1.equals(r2) ? Map.of(r1, EventGraph.union(s1, s2)) : Map.of(r1, s1, r2, s2);
-    }
-}
